Prep for v0.2
ADDED error handling for Broken pipe on send_message()

ADDED reconnect() function with a wait time of 1/4 seconds

ADDED recentException on data recv to reconnect on silent disconnect
server side (This is a dirty hack. Find a proper solution in the future)

REMOVED trueCount. It was no longer needed once the auto-reconnect hack
was implemented

FIXED disconnect(). Socket shutdown required an argument?

FIXED KeyboardInterrupt. Small print() fix to move "Qutting" to a
newline